app/requirements.txt
Flask-RESTful==0.3.7
Flask==1.1.1
gunicorn==20.0.4
python-dotenv==0.19.0
flask-apispec==0.11.0
jinja2
requests
lxml==4.9.3
ncclient==0.6.13
urllib3==1.26.14
xmltodict==0.13.0
